[fix] Temporary solution for wrong entity deleting

+++ CactusCare.BLL/Services/DoctorService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -44,6 +45,9 @@ namespace CactusCare.BLL.Services
 
         public async Task DeleteAsync(int id)
         {
+            if ((await _unitOfWork.ReviewRepository.GetAllAsync()).Any(d => d.DoctorId.Equals(id)))
+                throw new ConstraintException();
+
             await _unitOfWork.DoctorRepository.DeleteAsync(id);
         }
     }
